pyproject.toml
[project]
name = "builderx"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.13.1",
    "python-dateutil>=2.9.0.post0",
    "uvicorn>=0.35.0",
]
